Consolidate notebook code and move python code into package. Delete trivial pdfs
