+++ notebook/skill2vec_1K.csv
